+++ source/components/theme-switcher/index.js
@@ -1,0 +1,2 @@
+import "./theme-switcher.less"
+import "./theme-switcher.js"